0-DataEngineer/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
flake8==7.1.0
greenlet==3.0.3
idna==3.7
mccabe==0.7.0
numpy==2.0.0
pandas==2.2.2
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.31
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
